requirements.txt
django
psycopg2
djangorestframework
markdown
django-filter
pillow
djangorestframework_simplejwt
django-cors-headers
django-storages
boto3
django_sendmail_backend
